Add files via upload
In this mini-project, we addressed the relationship between the coal power plant capacity and the distance to the nearest railway infrastructure. We built a linear regression model to assess the relationship, and the result showed that there is a negative correlation as distance to the railways does not affect the plant capacity.